feat(HU05): añadir repositorio IPerfilEstudiante y CRUD para PerfilEstudiante
Crear IPerfilEstudianteRepositorio extends JpaRepository<PerfilEstudiante, Long> con método findByEstudianteId(Long)
Añadir PerfilEstudianteServicios con operaciones CRUD y búsqueda por estudianteId
Añadir PerfilEstudianteControlador con endpoints REST (/perfiles, /perfiles/{id}, /perfiles/estudiante/{id})
Ajustar entidades relacionadas (id -> Long) e IEstudianteRepositorio a JpaRepository<Estudiante, Long>
Manejo de errores consistente usando MensajeError
Implementa HU05: permitir acceso y gestión de perfiles de estudiante.

+++ src/main/java/com/example/FrankySabado/modelos/Estudiante.java
@@ -11,9 +11,9 @@ import java.util.ArrayList;
 @Table(name = "estudiantes")
 public class Estudiante {
 
-    @Id
+        @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "promedio", nullable = false, unique = false)
     private Double promedio;
     @Column(name = "fechaNacimiento", nullable = false, unique = false)
@@ -40,17 +40,17 @@ public class Estudiante {
     public Estudiante() {
     }
 
-    public Estudiante(Integer id, Double promedio, LocalDate fechaNacimiento) {
+        public Estudiante(Long id, Double promedio, LocalDate fechaNacimiento) {
         this.id = id;
         this.promedio = promedio;
         this.fechaNacimiento = fechaNacimiento;
     }
 
-    public Integer getId() {
+        public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
